+++ statistics/skew_source.c
@@ -15,7 +15,7 @@ FUNCTION(gsl_stats,skew_with_mean_and_sd) (const BASE data[], const size_t n,
 {
   /* takes a dataset and finds the skewness */
 
-  double skew = 0;
+  long double skew = 0;
   size_t i;
 
   /* find the sum of the cubed deviations, normalized by the sd. */
@@ -25,8 +25,8 @@ FUNCTION(gsl_stats,skew_with_mean_and_sd) (const BASE data[], const size_t n,
 
   for (i = 0; i < n; i++)
     {
-      const double x = (data[i] - mean) / sd;
-      skew += (x * x * x - skew) / ((double)(i + 1));
+      const long double x = (data[i] - mean) / sd;
+      skew += (x * x * x - skew) / (i + 1);
     }
 
   return skew;
